Extract the PublicationManifest
