+++ r2-shared-swiftTests/Publication/PublicationTests.swift
@@ -14,7 +14,7 @@ import XCTest
 
 class PublicationTests: XCTestCase {
     
-    func testParseMinimalJSON() {
+    func testParseJSON() {
         XCTAssertEqual(
             try? Publication(json: [
                 "metadata": ["title": "Title"],
@@ -26,159 +26,23 @@ class PublicationTests: XCTestCase {
                 ]
             ]),
             Publication(
-                metadata: Metadata(title: "Title"),
-                links: [Link(href: "/manifest.json", rels: ["self"])],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")]
+                manifest: PublicationManifest(
+                    metadata: Metadata(title: "Title"),
+                    links: [Link(href: "/manifest.json", rels: ["self"])],
+                    readingOrder: [Link(href: "/chap1.html", type: "text/html")]
+                )
             )
         )
     }
-    
-    func testParseFullJSON() {
-        XCTAssertEqual(
-            try? Publication(json: [
-                "@context": "https://readium.org/webpub-manifest/context.jsonld",
-                "metadata": ["title": "Title"],
-                "links": [
-                    ["href": "/manifest.json", "rel": "self"]
-                ],
-                "readingOrder": [
-                    ["href": "/chap1.html", "type": "text/html"]
-                ],
-                "resources": [
-                    ["href": "/image.png", "type": "image/png"]
-                ],
-                "toc": [
-                    ["href": "/cover.html"],
-                    ["href": "/chap1.html"]
-                ],
-                "sub": [
-                    "links": [
-                        ["href": "/sublink"]
-                    ]
-                ]
-            ]),
-            Publication(
-                context: ["https://readium.org/webpub-manifest/context.jsonld"],
-                metadata: Metadata(title: "Title"),
-                links: [Link(href: "/manifest.json", rels: ["self"])],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")],
-                resources: [Link(href: "/image.png", type: "image/png")],
-                tableOfContents: [Link(href: "/cover.html"), Link(href: "/chap1.html")],
-                otherCollections: [PublicationCollection(role: "sub", links: [Link(href: "/sublink")])]
-            )
-        )
-    }
-    
-    func testParseContextAsArray() {
-        XCTAssertEqual(
-            try? Publication(json: [
-                "@context": ["context1", "context2"],
-                "metadata": ["title": "Title"],
-                "links": [
-                    ["href": "/manifest.json", "rel": "self"]
-                ],
-                "readingOrder": [
-                    ["href": "/chap1.html", "type": "text/html"]
-                ]
-                ]),
-            Publication(
-                context: ["context1", "context2"],
-                metadata: Metadata(title: "Title"),
-                links: [Link(href: "/manifest.json", rels: ["self"])],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")]
-            )
-        )
-    }
-    
-    func testParseInvalidJSON() {
-        XCTAssertThrowsError(try Publication(json: ""))
-    }
-    
-    func testParseJSONRequiresMetadata() {
-        XCTAssertThrowsError(try Publication(json: [
-            "links": [
-                ["href": "/manifest.json", "rel": "self"]
-            ],
-            "readingOrder": [
-                ["href": "/chap1.html", "type": "text/html"]
-            ]
-        ]))
-    }
-    
-    func testParseJSONSpineAsReadingOrder() {
-        // `readingOrder` used to be `spine`, so we parse `spine` as a fallback.
-        XCTAssertEqual(
-            try? Publication(json: [
-                "metadata": ["title": "Title"],
-                "links": [
-                    ["href": "/manifest.json", "rel": "self"]
-                ],
-                "spine": [
-                    ["href": "/chap1.html", "type": "text/html"]
-                ]
-            ]),
-            Publication(
-                metadata: Metadata(title: "Title"),
-                links: [Link(href: "/manifest.json", rels: ["self"])],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")]
-            )
-        )
-    }
-                    
-    func testParseJSONIgnoresReadingOrderWithoutType() {
-        XCTAssertEqual(
-            try Publication(json: [
-                "metadata": ["title": "Title"],
-                "links": [
-                    ["href": "/manifest.json", "rel": "self"]
-                ],
-                "readingOrder": [
-                    ["href": "/chap1.html", "type": "text/html"],
-                    ["href": "/chap2.html"]
-                ]
-            ]),
-            Publication(
-                metadata: Metadata(title: "Title"),
-                links: [
-                    Link(href: "/manifest.json", rels: ["self"]),
-                ],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")]
-            )
-        )
-    }
-    
-    func testParseJSONIgnoresRessourcesWithoutType() {
-        XCTAssertEqual(
-            try Publication(json: [
-                "metadata": ["title": "Title"],
-                "links": [
-                    ["href": "/manifest.json", "rel": "self"]
-                ],
-                "readingOrder": [
-                    ["href": "/chap1.html", "type": "text/html"]
-                ],
-                "resources": [
-                    ["href": "/withtype", "type": "text/html"],
-                    ["href": "/withouttype"]
-                ]
-            ]),
-            Publication(
-                metadata: Metadata(title: "Title"),
-                links: [
-                    Link(href: "/manifest.json", rels: ["self"]),
-                ],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")],
-                resources: [Link(href: "/withtype", type: "text/html")]
-            )
-        )
-    }
-    
-    func testGetMinimalJSON() {
+
+    func testGetJSON() {
         AssertJSONEqual(
             Publication(
-                metadata: Metadata(title: "Title"),
-                links: [Link(href: "/manifest.json", rels: ["self"])],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")]
+                manifest: PublicationManifest(
+                    metadata: Metadata(title: "Title"),
+                    links: [Link(href: "/manifest.json", rels: ["self"])],
+                    readingOrder: [Link(href: "/chap1.html", type: "text/html")]
+                )
             ).json,
             [
                 "metadata": ["title": "Title", "readingProgression": "auto"],
@@ -191,57 +55,5 @@ class PublicationTests: XCTestCase {
             ]
         )
     }
-    
-    func testGetFullJSON() {
-        AssertJSONEqual(
-            Publication(
-                context: ["https://readium.org/webpub-manifest/context.jsonld"],
-                metadata: Metadata(title: "Title"),
-                links: [Link(href: "/manifest.json", rels: ["self"])],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")],
-                resources: [Link(href: "/image.png", type: "image/png")],
-                tableOfContents: [Link(href: "/cover.html"), Link(href: "/chap1.html")],
-                otherCollections: [PublicationCollection(role: "sub", links: [Link(href: "/sublink")])]
-            ).json,
-            [
-                "@context": ["https://readium.org/webpub-manifest/context.jsonld"],
-                "metadata": ["title": "Title", "readingProgression": "auto"],
-                "links": [
-                    ["href": "/manifest.json", "rel": ["self"], "templated": false]
-                ],
-                "readingOrder": [
-                    ["href": "/chap1.html", "type": "text/html", "templated": false]
-                ],
-                "resources": [
-                    ["href": "/image.png", "type": "image/png", "templated": false]
-                ],
-                "toc": [
-                    ["href": "/cover.html", "templated": false],
-                    ["href": "/chap1.html", "templated": false]
-                ],
-                "sub": [
-                    "links": [
-                        ["href": "/sublink", "templated": false]
-                    ]
-                ]
-            ]
-        )
-    }
-    
-    func testGetFullManifest() {
-        XCTAssertEqual(
-            Publication(
-                context: ["https://readium.org/webpub-manifest/context.jsonld"],
-                metadata: Metadata(title: "Title"),
-                links: [Link(href: "/manifest.json", rels: ["self"])],
-                readingOrder: [Link(href: "/chap1.html", type: "text/html")],
-                resources: [Link(href: "/image.png", type: "image/png")],
-                tableOfContents: [Link(href: "/cover.html"), Link(href: "/chap1.html")]
-            ).manifest,
-            """
-            {"@context":["https://readium.org/webpub-manifest/context.jsonld"],"links":[{"href":"/manifest.json","rel":["self"],"templated":false}],"metadata":{"readingProgression":"auto","title":"Title"},"readingOrder":[{"href":"/chap1.html","templated":false,"type":"text/html"}],"resources":[{"href":"/image.png","templated":false,"type":"image/png"}],"toc":[{"href":"/cover.html","templated":false},{"href":"/chap1.html","templated":false}]}
-            """.data(using: .utf8)
-        )
-    }
-    
+
 }
